backend/requirements.txt
flask
gunicorn
python-dotenv
mistralai
